+++ TDD/BlastAsia.DigiBook/BlastAsia.DigiBook.API/Controllers/EmployeesController.cs
@@ -47,18 +47,34 @@ namespace BlastAsia.DigiBook.API.Controllers
         public IActionResult CreateEmployee(
             [FromBody] Employee employee)
         {
-            var result = this.employeeService.Save(Guid.Empty, employee);
 
-            return CreatedAtAction("GetEmployee",
-                new { id = employee.EmployeeId }, result);
+            try
+            {
+                if (employee == null)
+                {
+                    return BadRequest();
+                }
+                var result = this.employeeService.Save(Guid.Empty, employee);
+
+                return CreatedAtAction("GetEmployees",
+                    new { id = employee.EmployeeId }, result);
+            }
+            catch (Exception exc)
+            {
+                return BadRequest(exc.Message);
+            }
         }
 
         [HttpDelete]
-        public IActionResult DeleteEmployeee(Guid id)
+        public IActionResult DeleteEmployee(Guid id)
         {
-
-            this.employeeRepository.Delete(id);
-            return NoContent();
+            var employeeToDelete = this.employeeRepository.Retrieve(id);
+            if (employeeToDelete != null)
+            {
+                this.employeeRepository.Delete(id);
+                return NoContent();
+            }
+            return NotFound();
         }
 
 
@@ -66,33 +82,59 @@ namespace BlastAsia.DigiBook.API.Controllers
         public IActionResult UpdateEmployee(
             [FromBody] Employee employee, Guid id)
         {
-            var existingEmployee = employeeRepository.Retrieve(id);
+            try
+            {
+                if (employee == null)
+                {
+                    return BadRequest();
+                }
 
-            existingEmployee.ApplyChanges(employee);
+                var existingEmployee = employeeRepository.Retrieve(id);
+                if (existingEmployee == null)
+                {
+                    return NotFound();
+                }
+                existingEmployee.ApplyChanges(employee);
 
-            this.employeeService.Save(id, existingEmployee);
-            return Ok(employee);
+                var result = this.employeeService.Save(id, employee);
+
+                return Ok(existingEmployee);
+            }
+            catch (Exception)
+            {
+                return BadRequest();
+            }
+
+            //this.employeeService.Save(id, existingEmployee);
+            //return Ok(employee);
         }
 
         [HttpPatch]
         public IActionResult PatchEmployee(
             [FromBody]JsonPatchDocument patchedEmployee, Guid id)
         {
-            if (patchedEmployee== null)
+            try
+            {
+                if (patchedEmployee == null)
+                {
+                    return BadRequest();
+                }
+
+                var employee = employeeRepository.Retrieve(id);
+                if (employee == null)
+                {
+                    return NotFound();
+                }
+
+                patchedEmployee.ApplyTo(employee);
+                employeeService.Save(id, employee);
+
+                return Ok(employee);
+            }
+            catch (Exception)
             {
                 return BadRequest();
             }
-
-            var employee = employeeRepository.Retrieve(id);
-            if (employee == null)
-            {
-                return NotFound();
-            }
-
-            patchedEmployee.ApplyTo(employee);
-            employeeService.Save(id, employee);
-
-            return Ok(employee);
         }
 
     }
